Notice Ing

+++ src/main/java/com/goodday/proj/api/shop/controller/ShopController.java
@@ -1,13 +1,17 @@
 package com.goodday.proj.api.shop.controller;
 
+import com.goodday.proj.api.constant.ErrorConst;
 import com.goodday.proj.api.file.FileStore;
 import com.goodday.proj.api.file.model.UploadFile;
+import com.goodday.proj.api.member.repository.MemberRepository;
 import com.goodday.proj.api.shop.dto.ProductFormDto;
 import com.goodday.proj.api.shop.model.Product;
 import com.goodday.proj.api.shop.repository.ShopRepository;
 import com.goodday.proj.api.shop.service.ShopService;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
@@ -22,6 +26,7 @@ public class ShopController {
 
     private final ShopService shopService;
     private final ShopRepository shopRepository;
+    private final MemberRepository memberRepository;
     private final FileStore fileStore;
 
     /**
@@ -31,7 +36,7 @@ public class ShopController {
      */
     @GetMapping
     public Map<String, Object> product(@RequestParam(required = false) Integer currentPage) {
-        if (currentPage == null || currentPage < 0) {
+        if (currentPage == null || currentPage < 1) {
             currentPage = 1;
         }
         return shopService.pageAndProductList(currentPage);
@@ -39,15 +44,25 @@ public class ShopController {
 
     /**
      * 상품 등록
+     *
      * @param form
      * @throws IOException
      */
     @PostMapping("/write")
-    public void writeProduct(@ModelAttribute ProductFormDto form) throws IOException {
+    public void writeProduct(@Valid @ModelAttribute ProductFormDto form, BindingResult bindingResult) throws IOException {
+        if (memberRepository.findSessionMemberByNo(form.getMemberNo()).get().getAdmin().equals("N")) {
+            throw new RuntimeException(ErrorConst.authError);
+        }
+        if (bindingResult.hasErrors()) {
+            throw new IllegalArgumentException(ErrorConst.bindingError);
+        }
         UploadFile thumbnail = fileStore.storeFile(form.getThumbnail());
         List<UploadFile> images = fileStore.storeFiles(form.getImages());
 
-        shopService.writeProduct(thumbnail, images, form);
+        int result = shopService.writeProduct(thumbnail, images, form);
+        if (result == 0) {
+            throw new RuntimeException(ErrorConst.insertError);
+        }
     }
 
     /**
@@ -61,35 +76,48 @@ public class ShopController {
     }
 
     /**
-     * 상품 수정 view
+     * 상품 수정 view 이동
      * @param proNo
      * @return Product
      */
-    @GetMapping("/edit/{proNo}")
+    @PostMapping("/{proNo}")
     public Product editProduct(@PathVariable Long proNo) {
         return shopRepository.findByNo(proNo);
     }
 
     /**
      * 상품 수정
+     *
      * @param proNo
      * @param form
      * @throws IOException
      */
-    @PatchMapping("/{proNo}")
-    public void updateProduct(@PathVariable Long proNo, @ModelAttribute ProductFormDto form) throws IOException {
-        UploadFile thumbnail = fileStore.storeFile(form.getThumbnail());
-        List<UploadFile> images = fileStore.storeFiles(form.getImages());
+    @PostMapping("/{proNo}/edit")
+    public void editProduct(@PathVariable Long proNo, @Valid @ModelAttribute ProductFormDto form,
+                              BindingResult bindingResult) throws IOException {
+        if (memberRepository.findSessionMemberByNo(form.getMemberNo()).get().getAdmin().equals("N")) {
+            throw new RuntimeException(ErrorConst.authError);
+        }
+        if (bindingResult.hasErrors()) {
+            throw new IllegalArgumentException(ErrorConst.bindingError);
+        }
 
-        shopService.writeProduct(thumbnail, images, form);
+        int result = shopService.editProductAndFile(proNo, form);
+        if (result == 0) {
+            throw new RuntimeException(ErrorConst.updateError);
+        }
     }
 
     /**
      * 상품 삭제
+     *
      * @param proNo
      */
     @DeleteMapping("/{proNo}")
-    public void deleteProduct(@PathVariable Long proNo) {
+    public void deleteProduct(@PathVariable Long proNo, @RequestParam Long memberNo) {
+        if (memberRepository.findSessionMemberByNo(memberNo).get().getAdmin().equals("N")) {
+            throw new RuntimeException(ErrorConst.authError);
+        }
         Product product = shopRepository.findByNo(proNo);
         fileStore.deleteFile(product.getThumbnail().getStoreFileName());
         product.getImages().stream().forEach(uploadFile -> fileStore.deleteFile(uploadFile.getStoreFileName()));
